catch-all/06_bots_telegram/04_clima_bot/requirements.txt
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
decorator==5.1.1
idna==3.7
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
tornado==6.4.1
urllib3~=1.26
